+++ ft_strncpy.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strncpy(char *dest, const char *src, unsigned int n)
+#include "libft.h"
+
+char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	unsigned int		i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && src[i])
